chore: move `.md` files to published package (#232)
This moves `README.md`, `LICENSE.md` and `CHANGELOG.md` from the
repository root into the `./addon` directory.

In doing so, the three files will be picked up by `npm pack` and
`npm publish` respectively, so that the npmjs.com listing shows
the proper `README.md` instead of an error message, and so that these
files are also available in the users `node_modules`.

It also adds symlinks from the old location to the new location, so that
GitHub still renders the repository overview correctly. It's important,
that the files are physically moved instead of just symlinked into the
package, as `npm pack` doesn't follow symlinks.

`CONTRIBUTING.md` is intentionally not moved, as it's only relevant in
the context of the repository.